Section_060_-_Day_60_-_Advanced_-_Make_POST_Requests_with_Flask_and_HTML_Forms/flask-contact-form/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
email-validator==2.0.0